Ensure that plugins cannot bypass version check
This change addresses three issues related to plugin version checking:

1. It is possible for plugins to bypass version checking or do it wrong
   (as a matter of fact, the mjpeg fallback sets a bad example)

2. The current plugin version check violates the C++ ODR, i.e.
   it relies on undefined behaviors when the header used to compile
   the plugin and to compile the agent are not identical,

3. A major.minor numbering scheme is not ideal for ABI checks.
   In particular, it makes it difficult to react to an incompatibility
   that was detected post release.

[More info]

1. Make it impossible to bypass version checking

The current code depends on the plugin implementing the version check
correctly by calling PluginVersionIsCompatible. To make things worse,
the only publicly available example gets this wrong and uses an
ad-hoc version check, so anybody copy-pasting this code will get it
wrong.

It is more robust to do the version check in the agent before calling
any method in the plugin. It ensures that version checking cannot be
bypassed, is done consistently and generates consistent error messages.

As an indication that the aproach is robust, the new check correctly
refuses to load older plugins that use the old version checking method:

    spice-streaming-agent[5167]:
        error loading plugin [...].so: no version information

2. ODR-related problems

The C++ One Definition Rule (ODR) states that all translation units
must see the same definitions. In the current code, when we call
Agent::PluginVersionIsCompatible from the plugin, it is an ODR
violation as soon as we have made any change in the Agent class
compared to what the plugin was compiled against.

The current code therefore relies on implementation dependent knowlege
of how virtual functions are laid out in the vtable, and puts
unnecessary constraints on the changes allowed in the classes
(e.g. it's not allowed to put anything before some member functions)

Since the mechanism for detecting versions changes, we can also change
the numbering scheme (doing it at any other time would be problematic).
The previous numbering scheme was major.minor, which makes it difficult
to fixup a case where an old version would turn out to introduce a
subtle incompatibility. The scheme adopted here has an oldest
compatible version instead, so it does not have this problem.

3. Major.minor numbering scheme

The major.minor numbering scheme initially selected makes it harder
to fixes cases where an incompatibility was detected after release.

For example, the major.minor version checking assumes that agent 1.21
is compatible with plugins 1.21, 1.13 or 1.03. If later testing
shows that 1.13 actually introduced an incompatibility, you have to
special-case 1.13 in the compatibility check.

An approach that does not have this problem is to rely on incremented
version numbers, with a "current" and "oldest compatible" version
number. This is used for example by libtools [1].

Since the change required for #1 and #2  introduces an ABI break,
it is as good a time as any to also change the numbering scheme,
since changing it later would introduce another unnecessary ABI break.

[1] https://www.gnu.org/software/libtool/manual/html_node/Libtool-versioning.html