Implement version checking for plugins without violating ODR
The C++ One Definition Rule (ODR) states that all translation units
must see the same definitions. In the current code, when we call
Agent::PluginVersionIsCompatible from the plugin, it is an ODR
violation as soon as we have made any change in the Agent class
compared to what the plugin was compiled against.

GCC respects the Standard C++ ABI on most platforms, so I believe this
could be made to work despite the violation as long as we never
changed the order of declarations, etc. But the point of the ABI check
is that we should be able to change the agent interface in any way we
want and be safe. And technically, the safe cases would still be an
ODR violation that relies on knowledge of implementation details.

Another issue is that we don't want to have to rely on the plugins to
do the version checks. It is more robust to do it in the agent, where
it ensures that it is done in a consistent way and with consistent
error messages. As a matter of fact, the new check is robust against
older plugins and will not load them.

Since the mechanism for detecting versions changes, we can also change
the numbering scheme (doing it at any other time would be problematic).
The previous numbering scheme was major.minor, which makes it difficult
to fixup a case where an old version would turn out to introduce a
subtle incompatibility. The scheme adopted here has an oldest
compatible version instead, so it does not have this problem.