explicit instead of static registration for built-in plugins
The static registration (that is, having a static list of pointers to
static plugin variables and calling a generic function in the main
initialization code path to register them) allows to add plugins without
registering each of them explicitly.

However, in a single codebase, and having very few plugins, there is
very little advantage to it and the tradeoff for the complexity of this
initialization is not worth it. A single call for every built-in plugin
is much more simple and clear.

Signed-off-by: Lukáš Hrázký <lhrazky@redhat.com>
Reviewed-by: Christophe de Dinechin <dinechin@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>