Get rid of C-style memset initializations, use C++ style aggregates
C++-style aggregate initialization with explicit field names is more
readable and requires less repeated typing.

Like zero initialization, aggregate initialization preserves C++ type
safety.  However, unlike zero initialization, it does not guarantee
that padding is properly initialized. This is not a real problem in
our code where all padding is explicit, but it is worth keeping in
mind.

Also note that GCC is a bit picky and will generate strange
"unsupported" or "not implemented" errors if the fields are not all
initialized in exact declaration order. It is useful, however, to get
an error if a field is added and not initialized.

Writing this patch also highlighted an inconsistency between the type
of a 'codec' local variable and the type used in the actual data
type. Changed the type of the variable to uint8_t to match that in the
structure declaration.