Get rid of C-style memset initializations, use C++ style aggregates
C++-style aggregate initialization with explicit field names is more
readable and requires less repeated typing.

Like zero initialization, aggregate initialization preserves C++ type
safety.  However, unlike zero initialization, it does not guarantee
that padding is properly initialized. This is not a real problem in
our code where all padding is explicit, but it is worth keeping in
mind.

Also note that designated initializers is a C99 or C++20 feature.
It is presently accepted by both gcc and clang, but not VS2015.
Also note that gcc is picky relative to clang, but closer to the
current working draft for C++20, in that it requires fields to be
initialized in declaration order and to all be initialized, and that
you don't get a clear error message but an "unsupported" or
"unimplementd" message if you do not follow that rule.

Writing this patch also highlighted an inconsistency between the type
of a 'codec' local variable and the type used in the actual data
type (which must obey C rules, since it's in the protocol).
In order to avoid a warning, this patch changes the type of the
variable to uint8_t to match that in the structure declaration.