Catch all std::exception derivatives
std::bad_alloc is not an std::runtime_exception, it was not caught before.
We do not catch(...) on purpose, since calling 'terminate' and the abort()
that results is probably the best course of action in case some unknown
exception is thrown, if only because we can get a core dump.