Implement handling of error messages from the server
Log error messages from the server to syslog (not aborting the agent).
Limits the messages to 1023 bytes, error messages from the server are
not expected to be longer. In the unlikely case the message is longer,
log the first 1023 bytes and throw an exception (because we don't read
out the rest of the message, causing desynchronization between the
server and the streaming agent).

Signed-off-by: Lukáš Hrázký <lhrazky@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>