Add exception handling classes
Throwing 'runtime_error' directly should be reserved for the support
library.  Add an 'Error' class as a base class for all errors thrown
by the streaming agent, as well as subclasses used to discriminate
between categories of error.

The Error class offers:
- a 'format_message' member function that formats a message without
  allocating memory, so that there is no risk of throwing another
  exception at throw time.

- a 'syslog' member function that sends the formatted message to the syslog.

The derived classes are:

- IOError encapsulates typical I/O errors that report errors using
  errno.  The format_message in that case prints the original message,
  along with the strerror message.

- The ReadError and WriteError are two classes derived from IOError,
  with the sole purpose of making it possible to discriminate the
  error type at catch time, e.g. to retry writes.

These classes are used in the subsequent patches