Properly join cursor thread on destruction
The code was calling thread.detach(), which made the thread
independant from the thread object. Presumably, that was done so that
attempting to destroy the thread object would not result in
'terminate()', but the result is that we could also attempt to close
the X11 connexion while the thread was still running, resulting in
an abort.

Instead, we use the quit_requested() to make sure that the thread is
stopped in case of error, and we join the thread during the destructor
to make sure that it suspended its execution.