Correctly deal with SIGINT/SIGTERM on running agent
When the agent gets a signal, we need to kill two threads:
- The main thread, which may be in a system call.
- The X11 cursor thread if it's running, which may be either
  waiting for an X11 event or in a write system call.

SIGINT and SIGTERM signals are handled in ConcreteAgent::handle_interrupt
by setting a flag indicating that the agent must quit. The flag can be
tested using the quit_requested() method.

For any thread interrupted in a system call (poll, read or write), the
system calls returns with EINTR. ConcreteAgent::check_if_quitting()
after these system calls checks if we must quit, and if so, throws a
QuitResquested in the current thread.  Throwing an exception is
necessary because our read_all and write_all no longer return any
value.

For the X11 cursor thread, a signal during a write will happen like
for the main thread. We simply need a try/catch in that thread to
properly finish it instead of terminate()-ing the process. This also
deals with write errors, which would also have caused terminate() before.

The X11 cursor thread may also be waiting in XNextEvent(). In order to
get a regular opportunity to check if we must quit, add a timed
select() call on the X11 file descriptor, which allows us to check for
a request to quit every 0.5 second. That delay is good enough for a
human operator, without introducing any significant system load.