Make the 'quit' flag a part of the agent
Encapsulate the 'quit' flag in the agent, which is ultimately
responsible for being able to tell all components in the system that
it's time to quit.

Make the flag an std::atomic<bool> since it can be used from multiple
threads or from a signal handler.

Move the handle_interrupt and register_interrupts in the agent.