Don't warn for possible devirtualization optimization
The reported warning (or error in gitlab CI) ) is something like:

    ../include/spice-streaming-agent/errors.hpp:66:7: error: Declaring type 'class spice::streaming_agent::OpenError' final would enable devirtualization of 1 call [-Werror=suggest-final-types]
     class OpenError : public IOError
           ^~~~~~~~~

The compiler attempts to detect if it would be possible to make a
virtual call non-virtual by tagging a class as final. This would make
sense in a closed environment, if all sources were known. But since we
compile the agent and the plugins separately, the compiler gets it
wrong and suggest we mark final classes that we may want to use as
base classes.