Ensure that plugins cannot bypass version check
This change addresses two issues related to plugin version checking:

1. It is possible for plugins to bypass version checking or do it wrong
   (as a matter of fact, the mjpeg fallback sets a bad example)

2. The current plugin version check violates the C++ ODR, i.e.
   it relies on undefined behaviors when the header used to compile
   the plugin and to compile the agent are not identical,

[More info]

1. Make it impossible to bypass version checking

The current code depends on the plugin implementing the version check
correctly by calling PluginVersionIsCompatible. To make things worse,
the only publicly available example gets this wrong and uses an
ad-hoc version check, so anybody copy-pasting this code will get it
wrong.

It is more robust to do the version check in the agent before calling
any method in the plugin. It ensures that version checking cannot be
bypassed, is done consistently and generates consistent error messages.

As an indication that the aproach is robust, the new check correctly
refuses to load older plugins that use the old version checking method:

    spice-streaming-agent[5167]:
        error loading plugin [...].so: no version information

2. ODR-related problems

The C++ One Definition Rule (ODR) states that all translation units
must see the same definitions. In the current code, when we call
Agent::PluginVersionIsCompatible from the plugin, it is an ODR
violation as soon as we have made any change in the Agent class
compared to what the plugin was compiled against.

The current code therefore relies on implementation dependent knowlege
of how virtual functions are laid out in the vtable, and puts
unnecessary constraints on the changes allowed in the classes
(e.g. it's not allowed to put anything before some member functions)

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>