Move out the cursor-updating code into it's own class
Wraps the code in a functor which can then be used in std::thread. Keeps
the call to std::thread::detach() for now, as the thread actually cannot
be properly joined. It spends most of the time blocking on XNextEvent()
and there is no way of interrupting it.

It seems XNextEvent() is acutally using a socket, so a proper solution
might be to integrate this into the loop of the main process.

Signed-off-by: Lukáš Hrázký <lhrazky@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>