start streaming: check num_codecs
The server sends StreamMsgStartStop to tell spice-streaming-agent
to start streaming and a list of available codecs.

The first uint8_t is the number of codecs.
Each following uint8_t is a codec.

This patch checks that the number of codecs in the message, as
reported by the server, is not too large.

Signed-off-by: Uri Lublin <uril@redhat.com>
Acked-by: Jonathon Jongsma <jjongsma@redhat.com>