Make error.hpp and the Error class a public API
Moves IOError and its descendants to stream-port.hpp, as those are I/O
errors related to that. Makes the Error class a public base class for
exceptions thrown across the API boundary.

Note exception ABI conpatibility is not guaranteed between different
compilers, but it is compatible between gcc and clang so for these two
compilers it is safe.

Signed-off-by: Lukáš Hrázký <lhrazky@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>